Added debug file
The debug file will console log what request are being made and how long
they took. It does this by overriding the open method to get a reference
point of time and to store the url. This should only be used to debug,
but wouldn't be the worst thing if it made it into production.